cleanups and vertical line artifact fix (#12)
* fix typo to allow gl screenshots

* slightly non-zero "blank" alpha no longer needed

Seems like the issue was solved by the change to premultiplied alpha and/or the layer restructuring.

* don't add an offset to layer combine fragment

On Windows at certain canvas sizes, I was seeing strange vertical line artifacts.  Removing this offset fixes them.  I think it's not necessary, since the fragment will already be given coordinates in the center of the pixel?

The code can be cleaned up better but I wanted to be clear about the change.