fix fading sprites
I finally figured out why the fading sprites didn't look right.  After combining the sprite color PixelColor.rgb and the Color.rgb from the (sprite) call, it needs to be scaled by the #:a argument to (sprite).

I was wrong before in the comment about making Color.rgb into premultiplied alpha.  It's not being combined in the normal way, it's tinting the sprite pixel color.  That also answers why after adding it we have to cap it to PixelColor.a to maintain the premultiplied alpha invariant.  Hope that makes sense.